MyFunctionProject/requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
azure-common==1.1.28
azure-core==1.26.4
azure-functions==1.14.0
azure-identity==1.13.0
azure-keyvault-secrets==4.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
msal==1.22.0
msal-extensions==1.0.0
multidict==6.0.4
portalocker==2.7.0
pycparser==2.21
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
six==1.16.0
twilio==8.2.1
typing-extensions==4.5.0
urllib3==2.0.2
yarl==1.9.2
